FEATURE: new category setting to add "Unassigned" navigation menu
When a category is used exclusively for managing topic assignments it can
be very useful to list all the unassigned topics. This introduces a new
category setting that allows admins to opt for this new navigation item.

It works both on mobile and desktop.

This uses an amended nav item API, to experience this fully you need latest
versions of Discourse.