UX: Improve error message when assigning a PM to a user who doesn't have access to the PM
Currently if you try to assign a PM to a user who can't see it, you get
an error message that says "@{user} can't be assigned since they don't
belong to assigned allowed groups". This commit improves the message and
now it tells you the user doesn't have access to the PM and that you
need to invite the user first before you can assign them.

Signed-off-by: OsamaSayegh <asooomaasoooma90@gmail.com>