FIX: Add autofocus to users chooser on the assign modal (#148)
The users chooser on the assign modal is the primary input field on the modal, so it makes to shift focus to it when the modal is opened. This PR needs https://github.com/discourse/discourse/commit/0f807ba85b869fffc9560d97632d3ca7471e7fc4 in core to work properly, but it won't cause errors (won't have any effects) if it's deployed to a site that doesn't have that commit.