updated assign_allowed_on_groups admin setting
I had a bit of trouble finding how to enable this feature for a user so I could assign topics to the user. This is the setting, so I updated it to explain that it allows assigning to users as well as allows the user to assign topics to others.