DEV: Make the assign tab in the user menu link to the assign page (#380)
As of https://github.com/discourse/discourse/commit/12ebdf0ff0f742d5363527908a4a2721e41b00f2, it's possible to make tabs in the experimental user menu link to a page when they're clicked on when they're active. This commit makes the assign tab links to `/my/activity/assigned` when it's active.