FIX: Deactivate active assignments attached to deleted targets
Before the implementation of the `post_destroyed` event handler in
the plugin, soft deleting assigned posts/topics did not deactivate the
assignment.
This left sites which assigned and deleted posts/topics before the
introduction of the event handler with 'orphaned' assignments in the UI.

This change deactivates these orphaned assignments.