DEV: Fix flaky system tests
Why this change?

For some reason, toggling the topic admin menu via the button in the
footer is causing the test to be flaky. I don't have enough time to
figure out the root cause but we want to avoid these tests being flaky.
As such we'll use the topic admin menu in the timeline instead to close
the topic.