DEV: Set `capture_log: true` for flaky spec
Why this change?

The spec is flaky and it seems to be DB transaction related where after
signing in as a user, the failure screenshot shows that the user has not
been signed in.

What does this change do?

Set `capture_log: true` which will log all ActiveRecord DB statements
execute while running the spec.